+++ CHANGELOG.md
@@ -2,6 +2,22 @@
 
 All notable changes to this project will be documented in this file.
 
+## [0.1.2] - 2023-10-13
+
+### Bug Fixes
+
+- History listing ([#148](https://github.com/liftedinit/alberto/issues/148))
+
+## [0.1.1] - 2023-10-12
+
+### Bug Fixes
+
+- Hsm account creation regression ([#147](https://github.com/liftedinit/alberto/issues/147))
+
+### Features
+
+- Disable Hardware Authenticator identities for unsupported Networks ([#140](https://github.com/liftedinit/alberto/issues/140))
+
 ## [0.1.0] - 2023-10-02
 
 ### Bug Fixes